[CI] Stop uploading metrics via Grafana API endpoint (#557)
This changes makes it so we stop uploading commit data to our Grafana
Prometheus data source. Using the BigQuery plugin for Grafana instead
allows us to query our BigQuery dataset directly and visualize all of
our historical data in our dashboard. This means that BigQuery is now
our primary data source and we are no longer using our Prometheus data
source for the operational maturity dashboard.

Since we're not using prometheus as a data source anymore, we can stop
uploading data to it.