[CI] Add notifications for premerge buildbots (#568)
This is so that the appropriate people are getting notifications when
there are failures related to the premerge buildbots. Most of the time
no action is needed as build failures happen reasonably often and are
usually fixed quickly. Having the notifications (especially for missing
workers) helps ensure that the infra is not down/broken for a while.

Currently adding the premerge rotation, myself and Caroline. We might
eventually want to only have the premerge rotation on this, but as we
stabilize the infrastructure, having more people get notifications
should be helpful.