Fix ClangBuilder for running LNT on Windows (#575)
This PR adds support for running the LLVM test-suite on Windows
buildbots using ClangBuilder. The changes ensure that the correct
executable paths and file extensions are used when running Python and
LNT scripts on Windows.

This depends on llvm/llvm-project#155226 which adds llvm-lit.cmd wrapper
to make llvm-lit directory runnable on windows.