Add flang-arm64-windows-msvc-testsuite builder (#576)
This PR adds a new buildbot to run LLVM testsuite on windows. It adds a
new builder and worker pair. The testsuite will run after the first
stage build skipping the ninja check step as we cover that on other
buildbots. This will primarily be the builder to test that ensures LNT
and LLVM testsuite are running without regressions on windows. We will
later build on it to add support for test more as our Windows hardware
availability improves.

This depends on #575 and llvm/llvm-project#155226.