Insert debugify-report into the test suite build directory (#582)
As part of the DebugifyBuilder, we add cflag arguments to the test suite
build step containing the path of a report file, where any bugs detected
by debugify will be written. Debugify appends to the file rather than
replacing it, as we wish to accumulate the bugs found in all invocations
of Clang during the test suite build step to a single file. Currently
however, this file is not cleaned up after builds, meaning that we are
constantly appending bugs to the same file after each test run, rather
than creating a fresh file - this contaminates the results for future
builds, as bugs will never disappear from the list even if they no
longer exist in LLVM. This patch moves the report file into the test
suite build directory, which will always be cleaned before each test
run, preventing the results of one run from seeping into subsequent
runs.