[CI] Fix rounding when determining number of API batches to query
When querying the GitHub API, we get a runtime error if the batch
size is a multiple of the total number of commits. One extra batch
is queried for 0 commits, which then raises an exception when
trying to access the results. This change corrects the number of
batches to query for in those cases.