[test-suite] Install LNT using requirements file
Instead of "setup.py develop". Fixes https://github.com/llvm/llvm-lnt/issues/76.

The new process is tested in CI so we can catch problems before
they get to the buildbots.

The bots only use the client so I've used that specific requirement
file as suggested.

The requirement file has an entry ".". This means "install the package
defined in the current directory". "current" means the current working directory,
not the directory the requirement file is in.

This is handled by running the install command from within
the LNT checkout. LNT will still be installed in the sandbox as
expected.