[ProfCheck] Do Not Use External Shell
There have been test fidelity issues where tests have started failing in the
external shell while working fine on the internal shell. The original list
of tests that seemed to be failing is no longer failing with the internal
shell, so enable it for now to see how things go.