[CI] Get premerge-advisor service working correctly
The container did not have an exposed port in the deployment which prevented
anything from accessing the server hosted within. The service also was not
properly selecting for endpoints, which in itself also caused the service to
not be accessible. After these changes I have confirmed that the premerge
advisor is accessible inside the cluster.