[CI] Make Advisor Autoupdate Schema
This patch makes the premerge advisor server autoupdate the DB schema if we
ever end up changing it. We handle this by just dropping the existing table
as preserving all data for all time is completely unnecessary in our case.
Unittests have also been added to ensure that the setup_db function works as
it is supposed to.

Reviewers: cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/616