[CI] Add Library for Mapping Commit SHAs to Indexes
We need to know the index of commits to make some heuristics easier to
write, for example, are there any reported failures in the last five
commits. This patch adds a library that allows getting the index of a
commit assuming linear history. The plan is to use this to ensure the
failures database has commit indices rather than commit SHAs.

Reviewers: cmtice

Reviewed By: cmtice

Pull Request: https://github.com/llvm/llvm-zorg/pull/624