Skip scheduling sensetive lit tests on Linaro quick bots
Our quick bots run on a shared machine and use whatever CPU they
can get their hands on. This works well to keep utilisation high
but causes flaky failures in tests that assume set timeouts.

Such as the 3 lit tests I'm skipping here. These tests are run
on all platforms and have not had real failures in a very long
time, so 2 fewer bots running them isn't a problem.

The option used is:
https://llvm.org/docs/CommandGuide/lit.html#cmdoption-lit-filter-out

Note that it filters based on name, which does not include
the path. Otherwise I would have included "lnt/test/" in
the pattern.

The alternative is assigning cores to our quick bots, which
is difficult with the resources we have access to. Or we could
keep chasing the "perfect" timeout value in the tests. Which
is an endless quest because the system can choose to stall
a process almost forever if it likes.