Add extended factory function for TestSuiteBuilder factories. (#630)
Added an initial version of TestSuiteBuilder.getLlvmTestSuiteSteps()
that is used to prepare the build steps to retrieve, build and run the
LLVM TestSuite tests. This factory supports the CMake buildbot step,
synchronization and run the tests on the remote boards and could be used
together with the composite factories, such as
UnifiedTreeBuilder.getCmakeExBuildFactory().

This factory is planned as a replacement for the current factory later.