[CI] Make dispatch_job.py bail if pod is stuck pending
We have been running into some cases where a pod will be stuck pending
for 10+ hours. The buildbot should just bail if the pod can't schedule
in 30 minutes rather than waiting so long. We also need to fix the
underlying issue causing pods to not schedule, but it is good to be
resilient to this case too.