[CI] Use Grace Periods instead of PDBs
GKE apparently does not like to upgrade the cluster with how we were attempting
to use PDBs. Switch over to grace periods specified per pod, which is probably
closer to what we actually intend anyways. This might have the chance of killing
a still running job, but that should be pretty rare given the vast majority of
jobs complete in under an hour and that we have maintenance windows set to
only run jobs in the deep troughs of demand.