[CI] Filter explanations by test file at DB query time
Previously when trying to explain failures for either flaky tests or
tests failing at head, we would not filter by test file from the
database. This would result in a lot of results then being filtered by
message, which does contain the test name. This is mostly a performance
optimization as the test message in most cases will contain the name,
but could still be a correctness issue if two tests with different names
failed with the same message.

Reviewers: dwblaikie, Keenuts, cmtice

Reviewed By: dwblaikie

Pull Request: https://github.com/llvm/llvm-zorg/pull/689