[CI] Add debug logging for premerge advisor explanations
We sometimes want to be able to debug how the premerge advisor makes
explanations, especially in the case of false positive/negatives
(although only the latter has been observed in the wild before). This
proves hard because it is difficult to understand the system state at
the moment the explanation request was made.

This patch adds some logging so we can see all relevant information
(namly the relevant failures stored by the premerge advisor in the DB).

Reviewers: Keenuts, dwblaikie, cmtice

Reviewed By: dwblaikie

Pull Request: https://github.com/llvm/llvm-zorg/pull/690